formatted w/ black; replaced original find-ppr with _ns version
