+++ find-ppr.py
@@ -2,6 +2,10 @@ import pykd
 import argparse
 from enum import Enum, auto
 
+# Example bad chars, change to whatever
+BADCHARS = [0x00, 0x0A, 0x0D]
+# BADCHARS = [0x00,0x02,0x03,0x09,0x0A,0x0D,0x20,0x2E,0x2F]
+
 
 class Module:
     # 00400000 00465000   diskpls    (deferred)
@@ -32,16 +36,28 @@ class PopR32(Enum):
     edi = auto()
 
 
+def checkBadChars(bAddr):
+    for i in bAddr:
+        if i in BADCHARS:
+            return "--"
+    return "OK"
+
+
 def main(args):
     modules = pykd.dbgCommand("lm")
-
+    totalGadgets = 0  # This tracks all the total number of usable gadgets
+    modGadgetCount = {}  # This tracks the number of gadgets per module
     for mod_line in modules.splitlines():
         module = Module(mod_line)
 
         if module.name.lower() not in [mod.lower() for mod in args.modules]:
             continue
-
+        numGadgets = 0  # This is the number of gadgets found in this module
         print(f"[+] searching {module.name} for pop r32; pop r32; ret")
+        print("[+] BADCHARS: ", end="")
+        for i in BADCHARS:
+            print("\\x{:02X}".format(i), end="")
+        print()
 
         for pop1 in range(0x58, 0x60):
 
@@ -56,16 +72,44 @@ def main(args):
 
                 for addr in result.splitlines():
                     try:
-                        print(
-                            f"[+] {module.name}::{addr}: pop {PopR32(pop1).name}; pop {PopR32(pop2).name}; ret"
-                        )
+                        bAddr = int(addr, 16).to_bytes(4, "little")
+                        bcChk = checkBadChars(bAddr)
+                        bAddrEsc = ""  # This is the escaped string containing the little endian addr for shellcode output
+                        for b in bAddr:
+                            bAddrEsc += "\\x{:02X}".format(b)
+                        if args.showbc and bcChk == "--":
+                            print(
+                                f"[{bcChk}] {module.name}::{addr}: pop {PopR32(pop1).name}; pop {PopR32(pop2).name}; ret ; {bAddrEsc}"
+                            )
+                        elif bcChk == "OK":
+                            print(
+                                f"[{bcChk}] {module.name}::{addr}: pop {PopR32(pop1).name}; pop {PopR32(pop2).name}; ret ; {bAddrEsc}"
+                            )
+                            numGadgets = numGadgets + 1
                     except ValueError:
                         # not a valid pop r32
                         pass
+        print(f"[+] {module.name}: Found {numGadgets} usable gadgets!")
+        modGadgetCount[module.name] = numGadgets  # Add to the dict
+        totalGadgets = (
+            totalGadgets + numGadgets
+        )  # Increment total number of gadgets found
+    print("\n---- STATS ----")  # Print out all the stats
+    print(">> BADCHARS: ", end="")
+    for i in BADCHARS:
+        print("\\x{:02X}".format(i), end="")
+    print()
+    print(f">> Usable Gadgets Found: {totalGadgets}")
+    print(">> Module Gadget Counts")
+    for m, c in modGadgetCount.items():
+        print("   - {}: {} ".format(m, c))
 
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
+    parser.add_argument(
+        "-s", dest="showbc", help="Show addresses with bad chars", action="store_true"
+    )
     parser.add_argument(
         "modules",
         help="module name(s) to search for pop pop ret (ex: find-ppr.py libspp diskpls libpal)",
@@ -73,3 +117,4 @@ if __name__ == "__main__":
     )
     args = parser.parse_args()
     main(args)
+    print("Done!")
